Make --header the default, and rename the CLI args
Going forward, we will assume by default that the first line of
a CSV file contains header rows. If that's not the case, it must
be explicitly stated via --no-header-row or the #header=absent
fragment syntax. The old OFFSET 1 convention is still recoginzed
if CLI and fragment don't say anything about the header.

This addresses #9.